+++ test/PrimitiveInventorySpec.hs
@@ -45,6 +45,7 @@ spec = describe "MLF.Primitive.Inventory" $ do
           PrimitiveInventory.stringEndsWithPrimitiveName,
           PrimitiveInventory.stringAppendPrimitiveName,
           PrimitiveInventory.stringFromCharPrimitiveName,
+          PrimitiveInventory.stringToListPrimitiveName,
           PrimitiveInventory.stringDropPrimitiveName,
           PrimitiveInventory.stringTakePrimitiveName,
           PrimitiveInventory.stringSlicePrimitiveName,
@@ -86,6 +87,8 @@ spec = describe "MLF.Primitive.Inventory" $ do
           name `shouldBe` PrimitiveInventory.stringAppendPrimitiveName
         PrimitiveInventory.PrimitiveNativeStringFromChar ->
           name `shouldBe` PrimitiveInventory.stringFromCharPrimitiveName
+        PrimitiveInventory.PrimitiveNativeStringToList ->
+          name `shouldBe` PrimitiveInventory.stringToListPrimitiveName
         PrimitiveInventory.PrimitiveNativeStringDrop ->
           name `shouldBe` PrimitiveInventory.stringDropPrimitiveName
         PrimitiveInventory.PrimitiveNativeStringTake ->